--- scopes/react/component-highlighter/component-highlighter/index.ts
@@ -1,2 +1,0 @@
-export { HoverHighlighter } from './hover-highlighter';
-export type { HoverHighlighterProps } from './hover-highlighter';
