+++ scopes/react/component-highlighter/hover-highlighter/hover-highlighter.compositions.tsx
