+++ scopes/react/component-highlighter/hover-highlighter/hover-highlighter.spec.tsx
