+++ scopes/react/component-highlighter/hover-highlighter/index.ts
