+++ scopes/react/component-highlighter/index.ts
@@ -1,5 +1,8 @@
-export { HoverHighlighter as ComponentHighlighter } from './component-highlighter';
-export type { HoverHighlighterProps as ComponentHighlightProps } from './component-highlighter';
+export { HybridHighlighter as ComponentHighlighter } from './hybrid-highligher';
+export type { HybridHighlighterProps as ComponentHighlightProps } from './hybrid-highligher';
+
+export { HoverHighlighter } from './hover-highlighter';
+export type { HoverHighlighterProps } from './hover-highlighter';
 
 export { MultiHighlighter } from './multi-highlighter';
 export type { MultiHighlighterProps } from './multi-highlighter';
